Move editor directory to batch-spec/edit
