+++ client/web/src/enterprise/batches/batch-spec/edit/EditorForm.tsx
@@ -3,9 +3,9 @@ import React, { useContext } from 'react'
 import { ThemeProps } from '@sourcegraph/shared/src/theme'
 import { Panel } from '@sourcegraph/wildcard'
 
-import { EditorFeedbackPanel } from '../../create/editor/EditorFeedbackPanel'
-import { MonacoBatchSpecEditor } from '../../create/editor/MonacoBatchSpecEditor'
 import { BatchSpecContext } from '../BatchSpecContext'
+import { EditorFeedbackPanel } from './editor/EditorFeedbackPanel'
+import { MonacoBatchSpecEditor } from './editor/MonacoBatchSpecEditor'
 
 import { LibraryPane } from './library/LibraryPane'
 import { WorkspacesPreview } from './workspaces-preview/WorkspacesPreview'
