+++ client/web/src/enterprise/batches/batch-spec/edit/editor/MonacoBatchSpecEditor.module.scss
